lambda/requirements.txt
psycopg2-binary==2.9.7
boto3==1.28.25
requests==2.31.0
python-multipart==0.0.6
xmltodict==0.13.0
python-amazon-paapi==5.0.1
